backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
supabase==2.7.4
groq==0.11.0
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.5.2
Pillow==10.4.0
aiofiles==24.1.0